refactor: refactor job-listings page component
- create a custom table that extends DataTable component
- colocate all table related components under table dir
