+++ src/routes/Employer/JobListings/table/JobListingsTableToolbar.tsx
@@ -1,0 +1,46 @@
+import { X } from "lucide-react";
+import { Table } from "@tanstack/react-table";
+import { Button } from "@/components/ui/button";
+import { DataTableFacetedFilter } from "@/components/table/DataTableFacetedFilter";
+import { location } from "./data";
+import { Input } from "@/components/ui/input";
+
+interface DataTableToolbarProps<TData> {
+  table: Table<TData>;
+}
+
+const JobListingsTableToolbar = <TData,>({ table }: DataTableToolbarProps<TData>) => {
+  const isFiltered = table.getState().columnFilters.length > 0;
+
+  return (
+    <div className="flex items-center justify-between">
+      <div className="flex flex-1 items-center space-x-2">
+        <Input
+          placeholder="Filter jobs..."
+          value={(table.getColumn("title")?.getFilterValue() as string) ?? ""}
+          onChange={(event) => table.getColumn("title")?.setFilterValue(event.target.value)}
+          className="h-8 w-[150px] lg:w-[250px]"
+        />
+        {table.getColumn("location") && (
+          <DataTableFacetedFilter
+            column={table.getColumn("location")}
+            title="Location"
+            options={location}
+          />
+        )}
+        {isFiltered && (
+          <Button
+            variant="ghost"
+            onClick={() => table.resetColumnFilters()}
+            className="h-8 px-2 lg:px-3"
+          >
+            Reset
+            <X className="ml-2 h-4 w-4" />
+          </Button>
+        )}
+      </div>
+    </div>
+  );
+};
+
+export default JobListingsTableToolbar;
