+++ src/routes/Employer/JobListings/table/data.tsx
@@ -1,0 +1,29 @@
+import { MapPin } from "lucide-react";
+
+export const location = [
+  {
+    value: "Manila",
+    label: "Manila",
+    icon: MapPin,
+  },
+  {
+    value: "Makati City",
+    label: "Makati City",
+    icon: MapPin,
+  },
+  {
+    value: "Taguig City",
+    label: "Taguig City",
+    icon: MapPin,
+  },
+  {
+    value: "Pasig City",
+    label: "Pasig City",
+    icon: MapPin,
+  },
+  {
+    value: "Quezon City",
+    label: "Quezon City",
+    icon: MapPin,
+  },
+];
